contact update

+++ src/services/contacts.js
@@ -8,11 +8,12 @@ export const getContacts = async ({
     sortOrder = SORT_ORDER.ASC,
     sortBy = 'name',
     filter = {},
+    userId,
 }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
 
-    const contactsQuery = ContactCollection.find();
+    const contactsQuery = ContactCollection.find({ userId });
 
     if (filter.type) {
         contactsQuery.where('contactType').equals(filter.type);
@@ -23,7 +24,7 @@ export const getContacts = async ({
     }
 
     const [contactsCount, contacts] = await Promise.all([
-        ContactCollection.countDocuments(),
+        ContactCollection.countDocuments({ userId }),
         contactsQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder }).exec(),
     ]);
 
@@ -35,11 +36,12 @@ export const getContacts = async ({
     };
 };
 
-export const getContactById = id => ContactCollection.findOne({ _id: id });
+export const getContactById = (id, userId) => ContactCollection.findOne({ _id: id, userId });
 export const createContact = payload => ContactCollection.create(payload);
-export const updateContact = async (_id, payload, options = {}) => {
+
+export const updateContact = async (_id, payload,userId, options = {}) => {
     const { upsert } = options;
-    const rawResult = await ContactCollection.findOneAndUpdate({ _id }, payload, {
+    const rawResult = await ContactCollection.findOneAndUpdate({ _id, userId }, payload, {
         new: true,
         upsert,
         includeResultMetadata: true,
@@ -51,4 +53,4 @@ export const updateContact = async (_id, payload, options = {}) => {
     };
 };
 
-export const deleteContact = id => ContactCollection.findOneAndDelete({ _id: id });
+export const deleteContact = (id, userId) => ContactCollection.findOneAndDelete({ _id: id, userId });
